Crate sumo

This crate provides a Rust interface to different parts of the SUMO traffic simulator.

Modules

normalize
Transforms a raw::Network into a Network that’s easier to reason about.

raw
Parse SUMO networks from XML without making any simplifications or transformations. A subset of the structures and fields defined at https://sumo.dlr.de/docs/Networks/PlainXML.html are produced.

Structs

Connection

Edge

EdgeID

InternalEdge
See https://sumo.dlr.de/docs/Networks/SUMO_Road_Networks.html#internal_edges

InternalLane

InternalLaneID

Junction

Lane

LaneID

Network
A normalized form of a SUMO network. A raw::Network is a direct representation of a .net.xml file. That’s further simplified to produce this structure, which should be easier to work with. The transformations:

NodeID

Enums

Direction

VehicleClass